Change value in the database from type integer to type float

+++ db/migrate/20190424191154_create_multipliers.rb
@@ -2,7 +2,7 @@ class CreateMultipliers < ActiveRecord::Migration[5.2]
   def change
     create_table :multipliers do |t|
       t.string :name
-      t.integer :multiplier_value
+      t.decimal :multiplier_value
 
       t.timestamps
     end
